Initial commit
Uploading the initial set up of the skill. Notice that this version would be refused from being published on the amazon store.

Needed fixes:
Users may experience challenges invoking the invocation name due to inclusion of the term, “essilor” We advise selecting a different term to ensure users are able to invoke the skill.

he skill provides health-related information, but it does not include a disclaimer in the skill's detailed description that it is not a substitute for professional medical advice. Detailed Description of the skill must include a disclaimer stating that the skill is not a substitute for professional advice. Example: "The information in this skill is provided for informational and educational purposes only. It is not a substitute for professional medical advice, treatment or diagnosis. Call your doctor to receive medical advice. If you think you may have a medical emergency, please dial your local emergency response phone number".